# specs/openapi/crm.yaml
openapi: 3.1.0
info:
  title: CRM API
  version: 0.2.0
  description: >
    CRMAPI — drafted fresh, since CRM has no prior adapter counterpart in the
    Personal Loan Solution's Core Banking Adapter Layer. Flagged for the loan
    domain's architecture team to add CRMAdapter to Section 16.1. Follows the
    same conventions as PartyAPI/AccountAPI/GLPostingAPI/PaymentAPI (OpenAPI
    3.1, shared JSON Schema objects via $ref, Idempotency-Key on mutating
    operations, full error coverage), with one absolute constraint unique to
    this file: no request or response schema anywhere in it may include a
    Money field, under any circumstance. This is enforced structurally —
    service-case.schema.json (../schemas/service-case.schema.json) has no
    property anywhere that references money.schema.json, and this file never
    defines a local field that would reintroduce one. CRM has no path into
    GLPostingAPI at all; every operation here is either a pure write of
    non-monetary interaction/case/preference data or a pure read of status —
    never a balance mutation, and never a trigger for one.

    v0.2.0 (BREAKING from v0.1.0) adds the case-lifecycle, relationship-
    manager, and communication-preferences surface the CRM Codegen Agent's
    role requires — updateCase/closeCase/reopenCase, addCaseNote/
    listCaseNotes, assignRelationshipManager/getRelationshipManager,
    get/updateCommunicationPreferences — none of which existed before this
...
    getCustomer360.
...
servers:
...
security:
  - serviceAuth: []
paths:
  /interactions:log:
    post:
      operationId: logInteraction
      summary: Log a customer interaction event
      description: >
        Write path for the CRM-facing domain events published by other
        modules' adapters (loan.account.booked, loan.account.disbursed,
        loan.repayment.posted, delinquency.status.changed, loan.account.closed
        — see each event's x-consumers: [CRMAdapter, ...] in its own AsyncAPI
        file). Never a general-purpose note field for balance data — eventType
        is a closed enum matching exactly those events. Idempotent on the
        Idempotency-Key header (value = the source event's own idempotency
        key, e.g. the disbursementId or repaymentId that produced it): a
        retried call never logs a duplicate interaction.
      tags: [Interaction]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogInteractionRequest'
            examples:
              disbursement:
                summary: "Idempotency-Key = 3f2a9e2e-6b1a-4c9d-8b3a-1e2f3a4b5c6d"
                value:
                  loanAccountId: "8b3a1e2f-3a4b-5c6d-9e2e-6b1a4c9d3f2a"
                  eventType: ACCOUNT_DISBURSED
                  occurredAt: "2026-08-16T09:14:00Z"
      responses:
        '201':
          description: Interaction logged
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/Interaction'
        '200':
          description: Idempotent replay — original interaction returned, nothing re-logged
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases:open:
    post:
      operationId: openCase
      summary: Open a new service case
      description: >
        Human-initiated (Customer Service Rep), not triggered by a domain
        event — unlike logInteraction. Idempotent on the Idempotency-Key
        header (value = caseId): a retried call returns the original case
        rather than opening a duplicate.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenCaseRequest'
            examples:
              standard:
                summary: "Idempotency-Key = d4e5f6a7-3333-4a5b-9c8d-4a5b6c7d8e9f"
                value:
                  partyId: "9c8d2e3f-4a5b-6c7d-8b3a-1e2f3a4b5c6d"
                  loanAccountId: "8b3a1e2f-3a4b-5c6d-9e2e-6b1a4c9d3f2a"
                  reasonCode: PAYMENT_DISPUTE
                  assignedTo: "csr.jdoe"
      responses:
        '201':
          description: Case opened
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '200':
          description: Idempotent replay — original case returned, nothing re-opened
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: partyId (or loanAccountId, if supplied) does not resolve to an existing resource
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases/{caseId}:
    get:
      operationId: getCase
      summary: Retrieve a service case
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Current case state
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /customers/{partyId}/360:
    get:
      operationId: getCustomer360
      summary: Retrieve a consolidated, non-monetary view of a customer
      description: >
        Read model combining loan account status summaries (status only — no
        principal, balance, or term data), recent interactions, and open
        cases. This is the operation where the "no Money field under any
        circumstance" constraint is most tempting to violate (a natural urge
        to show "current balance" on a 360 view) and most important to hold —
        Customer360 is explicitly a CRM read model, not a ledger read model;
        a CSR needing balance data calls AccountAPI directly through a
        properly scoped, audited path, not through this endpoint.
        recentInteractions[].notes is PII-adjacent content — every call to
        this operation is access-logged (actor + timestamp), same as
        listCaseNotes.
      tags: [Customer360]
      parameters:
        - $ref: '#/components/parameters/PartyId'
      responses:
        '200':
          description: Consolidated customer view
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/Customer360'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases/{caseId}:update:
    post:
      operationId: updateCase
      summary: Update a case in progress — status transition and/or reassignment
      description: >
        Permits Open→InProgress, InProgress→Resolved, and Open→Resolved
        transitions, and/or changing assignedTo — never a transition to or
        from Closed (see closeCase/reopenCase, which own that boundary for
        clearer audit/event semantics). Requires expectedVersion (the
        caller's last-read ServiceCase.version) for optimistic concurrency:
        if another update has landed since, this returns 409 rather than
        silently overwriting it — the caller re-reads via getCase and
        retries. Idempotent on the Idempotency-Key header.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        '200':
          description: Case updated (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            expectedVersion is stale (concurrent update), the requested
            status transition is not permitted from the case's current
            status, or the Idempotency-Key was reused with a materially
            different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases/{caseId}:close:
    post:
      operationId: closeCase
      summary: Close a case
      description: >
        Permitted from Open, InProgress, or Resolved. Idempotent — closing
        an already-closed case is a no-op success, not an error, since a
        retried request must not fail just because it already succeeded
        once.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseCaseRequest'
      responses:
        '200':
          description: Case closed (or already closed — idempotent)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases/{caseId}:reopen:
    post:
      operationId: reopenCase
      summary: Reopen a closed case
      description: >
        Permitted ONLY from Closed — reopening any other status is a 409,
        not a no-op (unlike closeCase's idempotent-on-already-closed
        behavior: reopening a case that was never closed is a genuine
        state error, not a retry of a prior success). On reopen, status
        returns to Open (not InProgress), escalated resets to false, and
        slaDueAt is recomputed from now — a reopened case gets a fresh SLA
        clock, not the remainder of its original one.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenCaseRequest'
      responses:
        '200':
          description: Case reopened
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/ServiceCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Case is not currently Closed
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cases/{caseId}/notes:
    get:
      operationId: listCaseNotes
      summary: Retrieve every note on a case, oldest first
      description: >
        Append-only — a correction is a new note, never an edit of a prior
        one. CaseNote.body is PII-adjacent content: every call to this
        operation is access-logged (actor + timestamp) and this operation
        is excluded from any bulk/analytics export path.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Case notes, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      operationId: addCaseNote
      summary: Append a note to a case
      description: >
        Idempotent on the Idempotency-Key header. body is encrypted at rest
        (see internal/pii.Encryptor's counterpart in this service) and
        never appears in any domain event payload.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseNoteRequest'
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '200':
          description: Idempotent replay — original note returned, nothing re-added
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /customers/{partyId}/relationship-manager:
    get:
      operationId: getRelationshipManager
      summary: Retrieve a customer's current relationship-manager assignment
      description: >
        A persistent, per-customer relationship — distinct from any single
        case's assignedTo. relationshipManagerId is null if none is
        currently assigned.
      tags: [RelationshipManager]
      parameters:
        - $ref: '#/components/parameters/PartyId'
      responses:
        '200':
          description: Current assignment (relationshipManagerId null if none)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/RelationshipManagerAssignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      operationId: assignRelationshipManager
      summary: Assign (or reassign) a customer's relationship manager
      description: >
        Overwrites any prior assignment — the prior assignment is not
        deleted, it is superseded (retained in the audit trail, never a row
        deletion, same discipline as everywhere else in this system).
        Idempotent on the Idempotency-Key header.
      tags: [RelationshipManager]
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRelationshipManagerRequest'
      responses:
        '200':
          description: Relationship manager assigned (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/RelationshipManagerAssignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
  /customers/{partyId}/communication-preferences:
    get:
      operationId: getCommunicationPreferences
      summary: Retrieve a customer's communication preferences
      description: >
        Always returns 200 — a party that has never set preferences gets
        the conservative default (preferredChannel null, every opt-in
        false) rather than a 404, since every party implicitly has an
        (unset) preference state.
      tags: [CommunicationPreferences]
      parameters:
        - $ref: '#/components/parameters/PartyId'
      responses:
        '200':
          description: Current preferences (defaults if never explicitly set)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CommunicationPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      operationId: updateCommunicationPreferences
      summary: Replace a customer's communication preferences
      description: >
        Full replace, not a partial patch — every field in the request body
        is required so there is no ambiguity about what "not supplied"
        means for a preference toggle. Idempotent on the Idempotency-Key
        header.
      tags: [CommunicationPreferences]
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommunicationPreferencesRequest'
      responses:
        '200':
          description: Preferences updated (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CommunicationPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-supplied retry-safety token. Value convention documented per operation.
      schema:
        type: string
  securitySchemes:
    serviceAuth:
      type: oauth2
      description: OAuth2 client-credentials grant, internal service network + mTLS at the transport layer
      flows:
        clientCredentials:
          tokenUrl: https://auth.internal.core-banking.example/oauth2/token
          scopes:
            crm:interaction:write: Log interactions
            crm:case:write: Open, update, close, and reopen service cases; add case notes
            crm:case:read: Read service cases and case notes
            crm:customer360:read: Read the consolidated customer view
            crm:relationshipmanager:write: Assign relationship managers
            crm:relationshipmanager:read: Read relationship-manager assignments
            crm:commprefs:write: Update communication preferences
            crm:commprefs:read: Read communication preferences
...